Removed duplidation from TestFilters.java
